+++ CHANGELOG.rst
@@ -16,6 +16,12 @@ Change Log
 - [???] modeled dumps in grid2op (stuff that have a given energy max, and cannot produce more than the available energy)
 - [???] fix notebook 5 texts
 
+[0.9.2] - 2020-05-??
+---------------------
+- [ADDED] a function that returns the types of the action see `action.get_types()`
+- [ADDED] a class to "cache" the data in memory instead of reading it over an over again from disk (see
+  `grid2op.chronics.MultifolderWithCache`
+
 [0.9.1] - 2020-05-20
 ---------------------
 - [FIXED] a bug preventing to save gif with episode replay when there has been a game over before starting time step